+++ projects/swimlane/ngx-ui/src/lib/components/button-group/button-group.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ButtonGroupComponent } from './button-group.component';
+
+@NgModule({
+  declarations: [ButtonGroupComponent],
+  exports: [ButtonGroupComponent],
+  imports: [CommonModule]
+})
+export class ButtonGroupModule {}