+++ src/app/app.component.ts
@@ -236,6 +236,10 @@ export class AppComponent {
         {
           name: 'Dropzone',
           route: 'dropzone'
+        },
+        {
+          name: 'ButtonGroup',
+          route: 'button-group'
         }
       ]
     }